Not showing PHP errors (for PHP7) in production. Moved HTTP errors to a separate directory. Few more HTTP errors.
