+++ api/core/Directus/Exception/Http/InternalServerErrorException.php
@@ -1,0 +1,18 @@
+<?php
+
+/**
+ * Directus – <http://getdirectus.com>
+ *
+ * @link      The canonical repository – <https://github.com/directus/directus>
+ * @copyright Copyright 2006-2017 RANGER Studio, LLC – <http://rangerstudio.com>
+ * @license   GNU General Public License (v3) – <http://www.gnu.org/copyleft/gpl.html>
+ */
+
+namespace Directus\Exception\Http;
+
+use Directus\Exception\HttpException;
+
+class InternalServerErrorException extends HttpException
+{
+    protected $httpStatus = 500;
+}
