+++ api/routes/A1/Messages.php
@@ -11,7 +11,7 @@ use Directus\Database\TableGateway\DirectusMessagesRecipientsTableGateway;
 use Directus\Database\TableGateway\DirectusMessagesTableGateway;
 use Directus\Database\TableGateway\DirectusUsersTableGateway;
 use Directus\Database\TableGateway\RelationalTableGateway as TableGateway;
-use Directus\Exception\ForbiddenException;
+use Directus\Exception\Http\ForbiddenException;
 use Directus\Util\ArrayUtils;
 use Directus\Util\DateUtils;
 use Directus\View\JsonView;
